refactor(server): move Google Workspace logic out of server routes
